Context-Aware RAG System/requirements.txt
streamlit>=1.34.0
psycopg2-binary>=2.9.9
sentence-transformers>=2.7.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
